--- src/main/java/dku/capstone/CapstoneApplication.java
@@ -1,19 +1,0 @@
-package dku.capstone;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@SpringBootApplication
-public class CapstoneApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CapstoneApplication.class, args);
-    }
-
-    @GetMapping("/")
-    public String home() {
-        return "Hello, World!";
-    }
-
-}
