+++ src/test/java/dku_capstone/QRust/QRustApplicationTests.java
@@ -1,10 +1,10 @@
-package dku.capstone;
+package dku_capstone.QRust;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class CapstoneApplicationTests {
+class QRustApplicationTests {
 
     @Test
     void contextLoads() {
